@@ -4,18 +4,18 @@ publish_to: 'none'
 version: 1.0.0
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.0
-  custom_refresh_indicator: ^0.9.0
-  flutter_launcher_icons: ^0.8.1
-  grpc: ^2.3.0
-  protobuf: ^1.0.1
-  uuid: ^2.2.2
-  crayola: ^1.0.0
+  cupertino_icons: ^1.0.3
+  custom_refresh_indicator: ^1.0.0
+  flutter_launcher_icons: ^0.9.0
+  grpc: ^3.0.0
+  protobuf: ^2.0.0
+  uuid: ^3.0.4
+  crayola: ^2.0.0
 
 dev_dependencies:
   flutter_test:
