@@ -1,19 +1,13 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.25.0
+// 	protoc-gen-go v1.26.0
 // 	protoc        v3.13.0
 // source: api/v1/logomotive.proto
 
-package logomotive
+package v1
 
 import (
-	context "context"
-	proto "github.com/golang/protobuf/proto"
-	empty "github.com/golang/protobuf/ptypes/empty"
 	timestamp "github.com/golang/protobuf/ptypes/timestamp"
-	grpc "google.golang.org/grpc"
-	codes "google.golang.org/grpc/codes"
-	status "google.golang.org/grpc/status"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
@@ -26,10 +20,6 @@ const (
 	// Verify that runtime/protoimpl is sufficiently up-to-date.
 	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
 )
-
-// This is a compile-time assertion that a sufficiently up-to-date version
-// of the legacy proto package is being used.
-const _ = proto.ProtoPackageIsVersion4
 
 type LogEntry struct {
 	state         protoimpl.MessageState
@@ -102,16 +92,16 @@ func (x *LogEntry) GetTime() *timestamp.Timestamp {
 	return nil
 }
 
-type LabelList struct {
+type PushRequest struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	Labels []string `protobuf:"bytes,1,rep,name=labels,proto3" json:"labels,omitempty"`
+	Entry *LogEntry `protobuf:"bytes,1,opt,name=entry,proto3" json:"entry,omitempty"`
 }
 
-func (x *LabelList) Reset() {
-	*x = LabelList{}
+func (x *PushRequest) Reset() {
+	*x = PushRequest{}
 	if protoimpl.UnsafeEnabled {
 		mi := &file_api_v1_logomotive_proto_msgTypes[1]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
@@ -119,13 +109,13 @@ func (x *LabelList) Reset() {
 	}
 }
 
-func (x *LabelList) String() string {
+func (x *PushRequest) String() string {
 	return protoimpl.X.MessageStringOf(x)
 }
 
-func (*LabelList) ProtoMessage() {}
+func (*PushRequest) ProtoMessage() {}
 
-func (x *LabelList) ProtoReflect() protoreflect.Message {
+func (x *PushRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_api_v1_logomotive_proto_msgTypes[1]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
@@ -137,28 +127,28 @@ func (x *LabelList) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use LabelList.ProtoReflect.Descriptor instead.
-func (*LabelList) Descriptor() ([]byte, []int) {
+// Deprecated: Use PushRequest.ProtoReflect.Descriptor instead.
+func (*PushRequest) Descriptor() ([]byte, []int) {
 	return file_api_v1_logomotive_proto_rawDescGZIP(), []int{1}
 }
 
-func (x *LabelList) GetLabels() []string {
+func (x *PushRequest) GetEntry() *LogEntry {
 	if x != nil {
-		return x.Labels
+		return x.Entry
 	}
 	return nil
 }
 
-type TailQuery struct {
+type PushResponse struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	Label string `protobuf:"bytes,1,opt,name=label,proto3" json:"label,omitempty"`
+	Entry *LogEntry `protobuf:"bytes,1,opt,name=entry,proto3" json:"entry,omitempty"`
 }
 
-func (x *TailQuery) Reset() {
-	*x = TailQuery{}
+func (x *PushResponse) Reset() {
+	*x = PushResponse{}
 	if protoimpl.UnsafeEnabled {
 		mi := &file_api_v1_logomotive_proto_msgTypes[2]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
@@ -166,13 +156,13 @@ func (x *TailQuery) Reset() {
 	}
 }
 
-func (x *TailQuery) String() string {
+func (x *PushResponse) String() string {
 	return protoimpl.X.MessageStringOf(x)
 }
 
-func (*TailQuery) ProtoMessage() {}
+func (*PushResponse) ProtoMessage() {}
 
-func (x *TailQuery) ProtoReflect() protoreflect.Message {
+func (x *PushResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_api_v1_logomotive_proto_msgTypes[2]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
@@ -184,57 +174,348 @@ func (x *TailQuery) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use TailQuery.ProtoReflect.Descriptor instead.
-func (*TailQuery) Descriptor() ([]byte, []int) {
+// Deprecated: Use PushResponse.ProtoReflect.Descriptor instead.
+func (*PushResponse) Descriptor() ([]byte, []int) {
 	return file_api_v1_logomotive_proto_rawDescGZIP(), []int{2}
 }
 
-func (x *TailQuery) GetLabel() string {
+func (x *PushResponse) GetEntry() *LogEntry {
+	if x != nil {
+		return x.Entry
+	}
+	return nil
+}
+
+type FeedRequest struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Entry *LogEntry `protobuf:"bytes,1,opt,name=entry,proto3" json:"entry,omitempty"`
+}
+
+func (x *FeedRequest) Reset() {
+	*x = FeedRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_api_v1_logomotive_proto_msgTypes[3]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *FeedRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*FeedRequest) ProtoMessage() {}
+
+func (x *FeedRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_api_v1_logomotive_proto_msgTypes[3]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use FeedRequest.ProtoReflect.Descriptor instead.
+func (*FeedRequest) Descriptor() ([]byte, []int) {
+	return file_api_v1_logomotive_proto_rawDescGZIP(), []int{3}
+}
+
+func (x *FeedRequest) GetEntry() *LogEntry {
+	if x != nil {
+		return x.Entry
+	}
+	return nil
+}
+
+type FeedResponse struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Entry *LogEntry `protobuf:"bytes,1,opt,name=entry,proto3" json:"entry,omitempty"`
+}
+
+func (x *FeedResponse) Reset() {
+	*x = FeedResponse{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_api_v1_logomotive_proto_msgTypes[4]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *FeedResponse) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*FeedResponse) ProtoMessage() {}
+
+func (x *FeedResponse) ProtoReflect() protoreflect.Message {
+	mi := &file_api_v1_logomotive_proto_msgTypes[4]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use FeedResponse.ProtoReflect.Descriptor instead.
+func (*FeedResponse) Descriptor() ([]byte, []int) {
+	return file_api_v1_logomotive_proto_rawDescGZIP(), []int{4}
+}
+
+func (x *FeedResponse) GetEntry() *LogEntry {
+	if x != nil {
+		return x.Entry
+	}
+	return nil
+}
+
+type LabelsRequest struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+}
+
+func (x *LabelsRequest) Reset() {
+	*x = LabelsRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_api_v1_logomotive_proto_msgTypes[5]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *LabelsRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*LabelsRequest) ProtoMessage() {}
+
+func (x *LabelsRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_api_v1_logomotive_proto_msgTypes[5]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use LabelsRequest.ProtoReflect.Descriptor instead.
+func (*LabelsRequest) Descriptor() ([]byte, []int) {
+	return file_api_v1_logomotive_proto_rawDescGZIP(), []int{5}
+}
+
+type LabelsResponse struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Labels []string `protobuf:"bytes,1,rep,name=labels,proto3" json:"labels,omitempty"`
+}
+
+func (x *LabelsResponse) Reset() {
+	*x = LabelsResponse{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_api_v1_logomotive_proto_msgTypes[6]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *LabelsResponse) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*LabelsResponse) ProtoMessage() {}
+
+func (x *LabelsResponse) ProtoReflect() protoreflect.Message {
+	mi := &file_api_v1_logomotive_proto_msgTypes[6]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use LabelsResponse.ProtoReflect.Descriptor instead.
+func (*LabelsResponse) Descriptor() ([]byte, []int) {
+	return file_api_v1_logomotive_proto_rawDescGZIP(), []int{6}
+}
+
+func (x *LabelsResponse) GetLabels() []string {
+	if x != nil {
+		return x.Labels
+	}
+	return nil
+}
+
+type TailRequest struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Label string `protobuf:"bytes,1,opt,name=label,proto3" json:"label,omitempty"`
+}
+
+func (x *TailRequest) Reset() {
+	*x = TailRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_api_v1_logomotive_proto_msgTypes[7]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *TailRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*TailRequest) ProtoMessage() {}
+
+func (x *TailRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_api_v1_logomotive_proto_msgTypes[7]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use TailRequest.ProtoReflect.Descriptor instead.
+func (*TailRequest) Descriptor() ([]byte, []int) {
+	return file_api_v1_logomotive_proto_rawDescGZIP(), []int{7}
+}
+
+func (x *TailRequest) GetLabel() string {
 	if x != nil {
 		return x.Label
 	}
 	return ""
 }
 
+type TailResponse struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Entry *LogEntry `protobuf:"bytes,1,opt,name=entry,proto3" json:"entry,omitempty"`
+}
+
+func (x *TailResponse) Reset() {
+	*x = TailResponse{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_api_v1_logomotive_proto_msgTypes[8]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *TailResponse) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*TailResponse) ProtoMessage() {}
+
+func (x *TailResponse) ProtoReflect() protoreflect.Message {
+	mi := &file_api_v1_logomotive_proto_msgTypes[8]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use TailResponse.ProtoReflect.Descriptor instead.
+func (*TailResponse) Descriptor() ([]byte, []int) {
+	return file_api_v1_logomotive_proto_rawDescGZIP(), []int{8}
+}
+
+func (x *TailResponse) GetEntry() *LogEntry {
+	if x != nil {
+		return x.Entry
+	}
+	return nil
+}
+
 var File_api_v1_logomotive_proto protoreflect.FileDescriptor
 
 var file_api_v1_logomotive_proto_rawDesc = []byte{
 	0x0a, 0x17, 0x61, 0x70, 0x69, 0x2f, 0x76, 0x31, 0x2f, 0x6c, 0x6f, 0x67, 0x6f, 0x6d, 0x6f, 0x74,
-	0x69, 0x76, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x0a, 0x6c, 0x6f, 0x67, 0x6f, 0x6d,
-	0x6f, 0x74, 0x69, 0x76, 0x65, 0x1a, 0x1f, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x74, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70,
-	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x1b, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x65, 0x6d, 0x70, 0x74, 0x79, 0x2e, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x22, 0x7a, 0x0a, 0x08, 0x4c, 0x6f, 0x67, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x12,
-	0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x02, 0x69, 0x64, 0x12,
-	0x18, 0x0a, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09,
-	0x52, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x14, 0x0a, 0x05, 0x6c, 0x61, 0x62,
-	0x65, 0x6c, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x6c, 0x61, 0x62, 0x65, 0x6c, 0x12,
-	0x2e, 0x0a, 0x04, 0x74, 0x69, 0x6d, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e,
-	0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e,
-	0x54, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x04, 0x74, 0x69, 0x6d, 0x65, 0x22,
-	0x23, 0x0a, 0x09, 0x4c, 0x61, 0x62, 0x65, 0x6c, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x16, 0x0a, 0x06,
-	0x6c, 0x61, 0x62, 0x65, 0x6c, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x09, 0x52, 0x06, 0x6c, 0x61,
-	0x62, 0x65, 0x6c, 0x73, 0x22, 0x21, 0x0a, 0x09, 0x54, 0x61, 0x69, 0x6c, 0x51, 0x75, 0x65, 0x72,
-	0x79, 0x12, 0x14, 0x0a, 0x05, 0x6c, 0x61, 0x62, 0x65, 0x6c, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09,
-	0x52, 0x05, 0x6c, 0x61, 0x62, 0x65, 0x6c, 0x32, 0xe8, 0x01, 0x0a, 0x0a, 0x4c, 0x6f, 0x67, 0x6f,
-	0x6d, 0x6f, 0x74, 0x69, 0x76, 0x65, 0x12, 0x32, 0x0a, 0x04, 0x50, 0x75, 0x73, 0x68, 0x12, 0x14,
-	0x2e, 0x6c, 0x6f, 0x67, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x76, 0x65, 0x2e, 0x4c, 0x6f, 0x67, 0x45,
-	0x6e, 0x74, 0x72, 0x79, 0x1a, 0x14, 0x2e, 0x6c, 0x6f, 0x67, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x76,
-	0x65, 0x2e, 0x4c, 0x6f, 0x67, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x12, 0x36, 0x0a, 0x04, 0x46, 0x65,
-	0x65, 0x64, 0x12, 0x14, 0x2e, 0x6c, 0x6f, 0x67, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x76, 0x65, 0x2e,
-	0x4c, 0x6f, 0x67, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x1a, 0x14, 0x2e, 0x6c, 0x6f, 0x67, 0x6f, 0x6d,
-	0x6f, 0x74, 0x69, 0x76, 0x65, 0x2e, 0x4c, 0x6f, 0x67, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x28, 0x01,
-	0x30, 0x01, 0x12, 0x35, 0x0a, 0x04, 0x54, 0x61, 0x69, 0x6c, 0x12, 0x15, 0x2e, 0x6c, 0x6f, 0x67,
-	0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x76, 0x65, 0x2e, 0x54, 0x61, 0x69, 0x6c, 0x51, 0x75, 0x65, 0x72,
-	0x79, 0x1a, 0x14, 0x2e, 0x6c, 0x6f, 0x67, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x76, 0x65, 0x2e, 0x4c,
-	0x6f, 0x67, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x30, 0x01, 0x12, 0x37, 0x0a, 0x06, 0x4c, 0x61, 0x62,
-	0x65, 0x6c, 0x73, 0x12, 0x16, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x1a, 0x15, 0x2e, 0x6c, 0x6f,
-	0x67, 0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x76, 0x65, 0x2e, 0x4c, 0x61, 0x62, 0x65, 0x6c, 0x4c, 0x69,
-	0x73, 0x74, 0x42, 0x23, 0x5a, 0x21, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d,
+	0x69, 0x76, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x06, 0x61, 0x70, 0x69, 0x2e, 0x76,
+	0x31, 0x1a, 0x1f, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62,
+	0x75, 0x66, 0x2f, 0x74, 0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x2e, 0x70, 0x72, 0x6f,
+	0x74, 0x6f, 0x22, 0x7a, 0x0a, 0x08, 0x4c, 0x6f, 0x67, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x12, 0x0e,
+	0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x02, 0x69, 0x64, 0x12, 0x18,
+	0x0a, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52,
+	0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x14, 0x0a, 0x05, 0x6c, 0x61, 0x62, 0x65,
+	0x6c, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x6c, 0x61, 0x62, 0x65, 0x6c, 0x12, 0x2e,
+	0x0a, 0x04, 0x74, 0x69, 0x6d, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1a, 0x2e, 0x67,
+	0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x54,
+	0x69, 0x6d, 0x65, 0x73, 0x74, 0x61, 0x6d, 0x70, 0x52, 0x04, 0x74, 0x69, 0x6d, 0x65, 0x22, 0x35,
+	0x0a, 0x0b, 0x50, 0x75, 0x73, 0x68, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x26, 0x0a,
+	0x05, 0x65, 0x6e, 0x74, 0x72, 0x79, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x61,
+	0x70, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x4c, 0x6f, 0x67, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x52, 0x05,
+	0x65, 0x6e, 0x74, 0x72, 0x79, 0x22, 0x36, 0x0a, 0x0c, 0x50, 0x75, 0x73, 0x68, 0x52, 0x65, 0x73,
+	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x26, 0x0a, 0x05, 0x65, 0x6e, 0x74, 0x72, 0x79, 0x18, 0x01,
+	0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x4c, 0x6f,
+	0x67, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x52, 0x05, 0x65, 0x6e, 0x74, 0x72, 0x79, 0x22, 0x35, 0x0a,
+	0x0b, 0x46, 0x65, 0x65, 0x64, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x26, 0x0a, 0x05,
+	0x65, 0x6e, 0x74, 0x72, 0x79, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x61, 0x70,
+	0x69, 0x2e, 0x76, 0x31, 0x2e, 0x4c, 0x6f, 0x67, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x52, 0x05, 0x65,
+	0x6e, 0x74, 0x72, 0x79, 0x22, 0x36, 0x0a, 0x0c, 0x46, 0x65, 0x65, 0x64, 0x52, 0x65, 0x73, 0x70,
+	0x6f, 0x6e, 0x73, 0x65, 0x12, 0x26, 0x0a, 0x05, 0x65, 0x6e, 0x74, 0x72, 0x79, 0x18, 0x01, 0x20,
+	0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x4c, 0x6f, 0x67,
+	0x45, 0x6e, 0x74, 0x72, 0x79, 0x52, 0x05, 0x65, 0x6e, 0x74, 0x72, 0x79, 0x22, 0x0f, 0x0a, 0x0d,
+	0x4c, 0x61, 0x62, 0x65, 0x6c, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x22, 0x28, 0x0a,
+	0x0e, 0x4c, 0x61, 0x62, 0x65, 0x6c, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12,
+	0x16, 0x0a, 0x06, 0x6c, 0x61, 0x62, 0x65, 0x6c, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x09, 0x52,
+	0x06, 0x6c, 0x61, 0x62, 0x65, 0x6c, 0x73, 0x22, 0x23, 0x0a, 0x0b, 0x54, 0x61, 0x69, 0x6c, 0x52,
+	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x14, 0x0a, 0x05, 0x6c, 0x61, 0x62, 0x65, 0x6c, 0x18,
+	0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x6c, 0x61, 0x62, 0x65, 0x6c, 0x22, 0x36, 0x0a, 0x0c,
+	0x54, 0x61, 0x69, 0x6c, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x26, 0x0a, 0x05,
+	0x65, 0x6e, 0x74, 0x72, 0x79, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x10, 0x2e, 0x61, 0x70,
+	0x69, 0x2e, 0x76, 0x31, 0x2e, 0x4c, 0x6f, 0x67, 0x45, 0x6e, 0x74, 0x72, 0x79, 0x52, 0x05, 0x65,
+	0x6e, 0x74, 0x72, 0x79, 0x32, 0xeb, 0x01, 0x0a, 0x11, 0x4c, 0x6f, 0x67, 0x6f, 0x6d, 0x6f, 0x74,
+	0x69, 0x76, 0x65, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x12, 0x31, 0x0a, 0x04, 0x50, 0x75,
+	0x73, 0x68, 0x12, 0x13, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x50, 0x75, 0x73, 0x68,
+	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x14, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x76, 0x31,
+	0x2e, 0x50, 0x75, 0x73, 0x68, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x35, 0x0a,
+	0x04, 0x46, 0x65, 0x65, 0x64, 0x12, 0x13, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x46,
+	0x65, 0x65, 0x64, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x14, 0x2e, 0x61, 0x70, 0x69,
+	0x2e, 0x76, 0x31, 0x2e, 0x46, 0x65, 0x65, 0x64, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65,
+	0x28, 0x01, 0x30, 0x01, 0x12, 0x33, 0x0a, 0x04, 0x54, 0x61, 0x69, 0x6c, 0x12, 0x13, 0x2e, 0x61,
+	0x70, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x54, 0x61, 0x69, 0x6c, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
+	0x74, 0x1a, 0x14, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x54, 0x61, 0x69, 0x6c, 0x52,
+	0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x30, 0x01, 0x12, 0x37, 0x0a, 0x06, 0x4c, 0x61, 0x62,
+	0x65, 0x6c, 0x73, 0x12, 0x15, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x4c, 0x61, 0x62,
+	0x65, 0x6c, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x16, 0x2e, 0x61, 0x70, 0x69,
+	0x2e, 0x76, 0x31, 0x2e, 0x4c, 0x61, 0x62, 0x65, 0x6c, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e,
+	0x73, 0x65, 0x42, 0x2a, 0x5a, 0x28, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d,
 	0x2f, 0x61, 0x72, 0x6e, 0x65, 0x6d, 0x6f, 0x6c, 0x6c, 0x61, 0x6e, 0x64, 0x2f, 0x6c, 0x6f, 0x67,
-	0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x76, 0x65, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x6f, 0x6d, 0x6f, 0x74, 0x69, 0x76, 0x65, 0x2f, 0x61, 0x70, 0x69, 0x2f, 0x76, 0x31, 0x62, 0x06,
+	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -249,29 +530,39 @@ func file_api_v1_logomotive_proto_rawDescGZIP() []byte {
 	return file_api_v1_logomotive_proto_rawDescData
 }
 
-var file_api_v1_logomotive_proto_msgTypes = make([]protoimpl.MessageInfo, 3)
+var file_api_v1_logomotive_proto_msgTypes = make([]protoimpl.MessageInfo, 9)
 var file_api_v1_logomotive_proto_goTypes = []interface{}{
-	(*LogEntry)(nil),            // 0: logomotive.LogEntry
-	(*LabelList)(nil),           // 1: logomotive.LabelList
-	(*TailQuery)(nil),           // 2: logomotive.TailQuery
-	(*timestamp.Timestamp)(nil), // 3: google.protobuf.Timestamp
-	(*empty.Empty)(nil),         // 4: google.protobuf.Empty
+	(*LogEntry)(nil),            // 0: api.v1.LogEntry
+	(*PushRequest)(nil),         // 1: api.v1.PushRequest
+	(*PushResponse)(nil),        // 2: api.v1.PushResponse
+	(*FeedRequest)(nil),         // 3: api.v1.FeedRequest
+	(*FeedResponse)(nil),        // 4: api.v1.FeedResponse
+	(*LabelsRequest)(nil),       // 5: api.v1.LabelsRequest
+	(*LabelsResponse)(nil),      // 6: api.v1.LabelsResponse
+	(*TailRequest)(nil),         // 7: api.v1.TailRequest
+	(*TailResponse)(nil),        // 8: api.v1.TailResponse
+	(*timestamp.Timestamp)(nil), // 9: google.protobuf.Timestamp
 }
 var file_api_v1_logomotive_proto_depIdxs = []int32{
-	3, // 0: logomotive.LogEntry.time:type_name -> google.protobuf.Timestamp
-	0, // 1: logomotive.Logomotive.Push:input_type -> logomotive.LogEntry
-	0, // 2: logomotive.Logomotive.Feed:input_type -> logomotive.LogEntry
-	2, // 3: logomotive.Logomotive.Tail:input_type -> logomotive.TailQuery
-	4, // 4: logomotive.Logomotive.Labels:input_type -> google.protobuf.Empty
-	0, // 5: logomotive.Logomotive.Push:output_type -> logomotive.LogEntry
-	0, // 6: logomotive.Logomotive.Feed:output_type -> logomotive.LogEntry
-	0, // 7: logomotive.Logomotive.Tail:output_type -> logomotive.LogEntry
-	1, // 8: logomotive.Logomotive.Labels:output_type -> logomotive.LabelList
-	5, // [5:9] is the sub-list for method output_type
-	1, // [1:5] is the sub-list for method input_type
-	1, // [1:1] is the sub-list for extension type_name
-	1, // [1:1] is the sub-list for extension extendee
-	0, // [0:1] is the sub-list for field type_name
+	9,  // 0: api.v1.LogEntry.time:type_name -> google.protobuf.Timestamp
+	0,  // 1: api.v1.PushRequest.entry:type_name -> api.v1.LogEntry
+	0,  // 2: api.v1.PushResponse.entry:type_name -> api.v1.LogEntry
+	0,  // 3: api.v1.FeedRequest.entry:type_name -> api.v1.LogEntry
+	0,  // 4: api.v1.FeedResponse.entry:type_name -> api.v1.LogEntry
+	0,  // 5: api.v1.TailResponse.entry:type_name -> api.v1.LogEntry
+	1,  // 6: api.v1.LogomotiveService.Push:input_type -> api.v1.PushRequest
+	3,  // 7: api.v1.LogomotiveService.Feed:input_type -> api.v1.FeedRequest
+	7,  // 8: api.v1.LogomotiveService.Tail:input_type -> api.v1.TailRequest
+	5,  // 9: api.v1.LogomotiveService.Labels:input_type -> api.v1.LabelsRequest
+	2,  // 10: api.v1.LogomotiveService.Push:output_type -> api.v1.PushResponse
+	4,  // 11: api.v1.LogomotiveService.Feed:output_type -> api.v1.FeedResponse
+	8,  // 12: api.v1.LogomotiveService.Tail:output_type -> api.v1.TailResponse
+	6,  // 13: api.v1.LogomotiveService.Labels:output_type -> api.v1.LabelsResponse
+	10, // [10:14] is the sub-list for method output_type
+	6,  // [6:10] is the sub-list for method input_type
+	6,  // [6:6] is the sub-list for extension type_name
+	6,  // [6:6] is the sub-list for extension extendee
+	0,  // [0:6] is the sub-list for field type_name
 }
 
 func init() { file_api_v1_logomotive_proto_init() }
@@ -293,7 +584,7 @@ func file_api_v1_logomotive_proto_init() {
 			}
 		}
 		file_api_v1_logomotive_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*LabelList); i {
+			switch v := v.(*PushRequest); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -305,7 +596,79 @@ func file_api_v1_logomotive_proto_init() {
 			}
 		}
 		file_api_v1_logomotive_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*TailQuery); i {
+			switch v := v.(*PushResponse); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_api_v1_logomotive_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*FeedRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_api_v1_logomotive_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*FeedResponse); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_api_v1_logomotive_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*LabelsRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_api_v1_logomotive_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*LabelsResponse); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_api_v1_logomotive_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*TailRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_api_v1_logomotive_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*TailResponse); i {
 			case 0:
 				return &v.state
 			case 1:
@@ -323,7 +686,7 @@ func file_api_v1_logomotive_proto_init() {
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_api_v1_logomotive_proto_rawDesc,
 			NumEnums:      0,
-			NumMessages:   3,
+			NumMessages:   9,
 			NumExtensions: 0,
 			NumServices:   1,
 		},
@@ -335,252 +698,4 @@ func file_api_v1_logomotive_proto_init() {
 	file_api_v1_logomotive_proto_rawDesc = nil
 	file_api_v1_logomotive_proto_goTypes = nil
 	file_api_v1_logomotive_proto_depIdxs = nil
-}
-
-// Reference imports to suppress errors if they are not otherwise used.
-var _ context.Context
-var _ grpc.ClientConnInterface
-
-// This is a compile-time assertion to ensure that this generated file
-// is compatible with the grpc package it is being compiled against.
-const _ = grpc.SupportPackageIsVersion6
-
-// LogomotiveClient is the client API for Logomotive service.
-//
-// For semantics around ctx use and closing/ending streaming RPCs, please refer to https://godoc.org/google.golang.org/grpc#ClientConn.NewStream.
-type LogomotiveClient interface {
-	Push(ctx context.Context, in *LogEntry, opts ...grpc.CallOption) (*LogEntry, error)
-	Feed(ctx context.Context, opts ...grpc.CallOption) (Logomotive_FeedClient, error)
-	Tail(ctx context.Context, in *TailQuery, opts ...grpc.CallOption) (Logomotive_TailClient, error)
-	Labels(ctx context.Context, in *empty.Empty, opts ...grpc.CallOption) (*LabelList, error)
-}
-
-type logomotiveClient struct {
-	cc grpc.ClientConnInterface
-}
-
-func NewLogomotiveClient(cc grpc.ClientConnInterface) LogomotiveClient {
-	return &logomotiveClient{cc}
-}
-
-func (c *logomotiveClient) Push(ctx context.Context, in *LogEntry, opts ...grpc.CallOption) (*LogEntry, error) {
-	out := new(LogEntry)
-	err := c.cc.Invoke(ctx, "/logomotive.Logomotive/Push", in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-func (c *logomotiveClient) Feed(ctx context.Context, opts ...grpc.CallOption) (Logomotive_FeedClient, error) {
-	stream, err := c.cc.NewStream(ctx, &_Logomotive_serviceDesc.Streams[0], "/logomotive.Logomotive/Feed", opts...)
-	if err != nil {
-		return nil, err
-	}
-	x := &logomotiveFeedClient{stream}
-	return x, nil
-}
-
-type Logomotive_FeedClient interface {
-	Send(*LogEntry) error
-	Recv() (*LogEntry, error)
-	grpc.ClientStream
-}
-
-type logomotiveFeedClient struct {
-	grpc.ClientStream
-}
-
-func (x *logomotiveFeedClient) Send(m *LogEntry) error {
-	return x.ClientStream.SendMsg(m)
-}
-
-func (x *logomotiveFeedClient) Recv() (*LogEntry, error) {
-	m := new(LogEntry)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-func (c *logomotiveClient) Tail(ctx context.Context, in *TailQuery, opts ...grpc.CallOption) (Logomotive_TailClient, error) {
-	stream, err := c.cc.NewStream(ctx, &_Logomotive_serviceDesc.Streams[1], "/logomotive.Logomotive/Tail", opts...)
-	if err != nil {
-		return nil, err
-	}
-	x := &logomotiveTailClient{stream}
-	if err := x.ClientStream.SendMsg(in); err != nil {
-		return nil, err
-	}
-	if err := x.ClientStream.CloseSend(); err != nil {
-		return nil, err
-	}
-	return x, nil
-}
-
-type Logomotive_TailClient interface {
-	Recv() (*LogEntry, error)
-	grpc.ClientStream
-}
-
-type logomotiveTailClient struct {
-	grpc.ClientStream
-}
-
-func (x *logomotiveTailClient) Recv() (*LogEntry, error) {
-	m := new(LogEntry)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-func (c *logomotiveClient) Labels(ctx context.Context, in *empty.Empty, opts ...grpc.CallOption) (*LabelList, error) {
-	out := new(LabelList)
-	err := c.cc.Invoke(ctx, "/logomotive.Logomotive/Labels", in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
-// LogomotiveServer is the server API for Logomotive service.
-type LogomotiveServer interface {
-	Push(context.Context, *LogEntry) (*LogEntry, error)
-	Feed(Logomotive_FeedServer) error
-	Tail(*TailQuery, Logomotive_TailServer) error
-	Labels(context.Context, *empty.Empty) (*LabelList, error)
-}
-
-// UnimplementedLogomotiveServer can be embedded to have forward compatible implementations.
-type UnimplementedLogomotiveServer struct {
-}
-
-func (*UnimplementedLogomotiveServer) Push(context.Context, *LogEntry) (*LogEntry, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method Push not implemented")
-}
-func (*UnimplementedLogomotiveServer) Feed(Logomotive_FeedServer) error {
-	return status.Errorf(codes.Unimplemented, "method Feed not implemented")
-}
-func (*UnimplementedLogomotiveServer) Tail(*TailQuery, Logomotive_TailServer) error {
-	return status.Errorf(codes.Unimplemented, "method Tail not implemented")
-}
-func (*UnimplementedLogomotiveServer) Labels(context.Context, *empty.Empty) (*LabelList, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method Labels not implemented")
-}
-
-func RegisterLogomotiveServer(s *grpc.Server, srv LogomotiveServer) {
-	s.RegisterService(&_Logomotive_serviceDesc, srv)
-}
-
-func _Logomotive_Push_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(LogEntry)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(LogomotiveServer).Push(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: "/logomotive.Logomotive/Push",
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(LogomotiveServer).Push(ctx, req.(*LogEntry))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-func _Logomotive_Feed_Handler(srv interface{}, stream grpc.ServerStream) error {
-	return srv.(LogomotiveServer).Feed(&logomotiveFeedServer{stream})
-}
-
-type Logomotive_FeedServer interface {
-	Send(*LogEntry) error
-	Recv() (*LogEntry, error)
-	grpc.ServerStream
-}
-
-type logomotiveFeedServer struct {
-	grpc.ServerStream
-}
-
-func (x *logomotiveFeedServer) Send(m *LogEntry) error {
-	return x.ServerStream.SendMsg(m)
-}
-
-func (x *logomotiveFeedServer) Recv() (*LogEntry, error) {
-	m := new(LogEntry)
-	if err := x.ServerStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-func _Logomotive_Tail_Handler(srv interface{}, stream grpc.ServerStream) error {
-	m := new(TailQuery)
-	if err := stream.RecvMsg(m); err != nil {
-		return err
-	}
-	return srv.(LogomotiveServer).Tail(m, &logomotiveTailServer{stream})
-}
-
-type Logomotive_TailServer interface {
-	Send(*LogEntry) error
-	grpc.ServerStream
-}
-
-type logomotiveTailServer struct {
-	grpc.ServerStream
-}
-
-func (x *logomotiveTailServer) Send(m *LogEntry) error {
-	return x.ServerStream.SendMsg(m)
-}
-
-func _Logomotive_Labels_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(empty.Empty)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(LogomotiveServer).Labels(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: "/logomotive.Logomotive/Labels",
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(LogomotiveServer).Labels(ctx, req.(*empty.Empty))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
-var _Logomotive_serviceDesc = grpc.ServiceDesc{
-	ServiceName: "logomotive.Logomotive",
-	HandlerType: (*LogomotiveServer)(nil),
-	Methods: []grpc.MethodDesc{
-		{
-			MethodName: "Push",
-			Handler:    _Logomotive_Push_Handler,
-		},
-		{
-			MethodName: "Labels",
-			Handler:    _Logomotive_Labels_Handler,
-		},
-	},
-	Streams: []grpc.StreamDesc{
-		{
-			StreamName:    "Feed",
-			Handler:       _Logomotive_Feed_Handler,
-			ServerStreams: true,
-			ClientStreams: true,
-		},
-		{
-			StreamName:    "Tail",
-			Handler:       _Logomotive_Tail_Handler,
-			ServerStreams: true,
-		},
-	},
-	Metadata: "api/v1/logomotive.proto",
 }
